// src/types.ts
export interface Branch {
  id: string;
  code: string;
  name: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  managerName: string;
  status: 'ACTIVE' | 'INACTIVE';
  openingDate: string;
  taxRate: number;
  gstin?: string;
  accessPassword?: string;
}